// src/app/models/user-account.model.ts
export class UserAccount {
    userAccountId: string = "";
    firstName: string = "";
    lastName: string = "";
    email: string = "";
    password: string = "";
    birthday: string = "";
    gender: string = "";
    mobileNumber: number = 0;
    aboutMe: string = "";
    profilePhotoPath: string = "";
    username: string = "";
}